GeneralDashboard gemaakt voor beter gebruik van dashboard
