+++ src/main/java/com/proj2/gui/DashboardController.java
@@ -15,9 +15,8 @@ import java.io.IOException;
 import java.net.URL;
 import java.util.ResourceBundle;
 
-public class DashboardController extends MainController implements Initializable
+public class DashboardController extends GeneralDashboard implements Initializable
 {
-    public Label nameUser;
     public Label pointsUser;
 
     public void nieuweRit(ActionEvent actionEvent)
@@ -26,41 +25,15 @@ public class DashboardController extends MainController implements Initializable
         changeSceneTo(stage, "/com/proj2/nieuwe-rit.fxml", new NieuweRitController());
     }
 
-    public void showLeaderboard(ActionEvent actionEvent)
-    {
-        Stage stage = IControllerInfo.getStage(actionEvent);
-        changeSceneTo(stage, "/com/proj2/leaderboard.fxml", new LeaderboardController());
-    }
-
     @FXML
     void toHistory(ActionEvent actionEvent) {
         Stage stage = IControllerInfo.getStage(actionEvent);
         changeSceneTo(stage, "/com/proj2/trip-history.fxml", new TripHistoryController());
     }
 
-
-    @FXML
-    void toChangeProfile(ActionEvent event) {
-        Stage stage = IControllerInfo.getStage(event);
-        changeSceneTo(stage, "/com/proj2/change-profile.fxml", new ChangeDataController());
-    }
-
-
-    @FXML
-    void toRewards(ActionEvent event) {
-        Stage stage = IControllerInfo.getStage(event);
-        changeSceneTo(stage, "/com/proj2/rewards.fxml", new RewardController());
-    }
-
-    public void logOutButton(ActionEvent actionEvent) {
-        getUser().setLoggedIn(false);
-        ((Stage) ((Node) actionEvent.getSource()).getScene().getWindow()).close();
-    }
-
-
     @Override
     public void initialize(URL url, ResourceBundle resourceBundle) {
-        nameUser.setText(getUser().getName());
+        getNameUser().setText(getUser().getName());
         pointsUser.setText(Integer.toString(((User) getUser()).getPoints()));
     }
 }
